Temp

+++ app/routes/color.py
@@ -1,21 +1,23 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
-from app.db import SessionLocal
+from app.db import SessionLocal, get_db
+from app.dependencies import get_current_user
 from app.models.color import Color
+from app.models.user import User
 from app.schemas.color import ColorCreate,ColorOut,ColorUpdate
 
 router = APIRouter(prefix="/colors", tags=["colors"])
 
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
 
 #CREATE COLOR ( GET İŞLEMLERİ HARİÇ HER CRUD İŞLEMİ İÇİN ADMİN ŞARTI OLACAK.. )
 @router.post("/", response_model=ColorOut)
-def color_create(payload: ColorCreate, db: Session = Depends(get_db)):
+def color_create(payload: ColorCreate, db: Session = Depends(get_db),current_user: User = Depends(get_current_user)):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="kullanıcı yetkisi yok"
+        )
+    
     new_color = Color(**payload.model_dump())
     db.add(new_color)
     db.commit()
@@ -39,10 +41,16 @@ def get_color_by_id(color_id: int, db: Session = Depends(get_db)):
 
 #DELETE COLOR
 @router.delete("/{color_id}",response_model=ColorOut)
-def delete_color_by_id(color_id: int,db:Session = Depends(get_db)):
+def delete_color_by_id(color_id: int,db:Session = Depends(get_db),current_user: User = Depends(get_current_user)):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="kullanıcı yetkisi yok"
+        )
+
     color = db.query(Color).filter(Color.id == color_id).first()
     if not color: 
-        raise HTTPException(status_code=404, detail="color not found to delete")
+        raise HTTPException(status_code=404, detail="silmek istediğin renk yok")
     
     db.delete(color)
     db.commit()
@@ -50,7 +58,13 @@ def delete_color_by_id(color_id: int,db:Session = Depends(get_db)):
 
 #UPDATE COLOR
 @router.put("/{color_id}", response_model=ColorOut)
-def update_color_by_id(color_id: int, payload: ColorUpdate, db: Session = Depends(get_db)):
+def update_color_by_id(color_id: int, payload: ColorUpdate, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
+    if not current_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="kullanıcı yetkisi yok"
+        )    
+    
     color = db.query(Color).filter(Color.id == color_id).first()
     if not color:
         raise HTTPException(status_code=404, detail="Size not found with that id")
